% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillaryFunctions_general.R
\name{ancillaryFunctions_general}
\alias{ancillaryFunctions_general}
\alias{getMODIStileNames}
\title{A set of ancillary functions}
\usage{
getMODIStileNames(roi)
}
\arguments{
\item{roi}{sf polygon object. a region of interest}
}
\description{
A set of ancillary functions
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
