% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillaryFunctions_oPeNDAP.R
\name{ancillaryFunctions_oPeNDAP}
\alias{ancillaryFunctions_oPeNDAP}
\alias{getOpenDAPvector}
\title{A set of ancillary functions to retrieve data using OpenDAP}
\usage{
getOpenDAPvector(OpenDAPUrl, variableName)
}
\description{
A set of ancillary functions to retrieve data using OpenDAP
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
