% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Wrapper to download several datasets}
\usage{
downloadData(df_to_dl, username = NULL, password = NULL,
  parallelDL = FALSE)
}
\description{
This function enables to download datasets, enventually parallelizing the download.
}
