% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraFunctions_importData.R
\name{extraFunctions_importData}
\alias{extraFunctions_importData}
\alias{importData_modis_vnp}
\title{A set of functions to import as raster data downloaded through the \code{getURL} functions}
\usage{
importData_modis_vnp(path_to_nc, var_name)
}
\arguments{
\item{path_to_nc}{string. Path to the data in .nc4 format}

\item{var_name}{string. Dimension name}

\item{roi}{sf POLYGON. The region of interest in EPSG 4326}
}
\description{
set of functions to import as raster data downloaded through the \code{getURL} functions
}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
