% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_era5.R, R/getData_gpm.R
\name{getData_gpm}
\alias{getData_gpm}
\alias{getData_era5}
\title{Download Global Precipitation Measurement time series data}
\usage{
getData_era5(timeRange = as.POSIXlt(c("2010-01-01 18:00:00",
  "2010-01-02 09:00:00")),
  roi = st_read("/home/ptaconet/r_react/getData/ROI_test.kml", quiet =
  T), dimensions = c("10m_u_component_of_wind",
  "10m_v_component_of_wind"), username = NULL, password = NULL,
  download = FALSE, destFolder = getwd(), ...)

getData_gpm(timeRange = as.Date(c("2017-01-01", "2017-01-30")),
  roi = st_read("/home/ptaconet/r_react/getData/ROI_test.kml", quiet =
  T), collection = "GPM_3IMERGDF.06",
  dimensions = c("precipitationCal"), OpenDAPXVector = NULL,
  OpenDAPYVector = NULL, roiSpatialIndexBound = NULL,
  username = NULL, password = NULL, download = FALSE,
  destFolder = getwd(), ...)
}
\arguments{
\item{timeRange}{Date(s) of interest. Mandatory. See Details for addition information on how to provide the dates.}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the GPM collection of interest.}

\item{username}{string. EarthData username}

\item{password}{string. EarthData password}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}

\item{collection}{The GPM collection of interest}

\item{OpenDAPXVector}{numeric vector. Optional. The OpenDAP X (longitude) dimension vector.}

\item{OpenDAPYVector}{numeric vector. Optional. The OpenDAP Y (latitude) dimension vector.}

\item{roiSpatialIndexBound}{numeric vector. Optional.}

\item{timeRange}{Date(s) of interest. Mandatory. See Details for addition information on how to provide the dates.}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}

\item{collection}{The GPM collection of interest}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the GPM collection of interest.}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}

\item{OpenDAPXVector}{numeric vector. Optional. The OpenDAP X (longitude) dimension vector.}

\item{OpenDAPYVector}{numeric vector. Optional. The OpenDAP Y (latitude) dimension vector.}

\item{roiSpatialIndexBound}{numeric vector. Optional.}

\item{username}{string. EarthData username}

\item{password}{string. EarthData password}
}
\value{
a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

If \code{download} is set to TRUE, the data are downloaded to \code{destFolder}

a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

If \code{download} is set to TRUE, the data are downloaded to \code{destFolder}
}
\description{
This function enables to retrieve URLs of GPM datasets for a given ROI and time frame, and eventually download the data

This function enables to retrieve URLs of GPM datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))}) or as a POSIXlt single time or time range (e.g. "2010-01-01 18:00:00") for the half-hourly collection (GPM_3IMERGHH.06). If POSIXlt, times must be in UTC.
Arguments \code{OpenDAPOpenDAPXVector}, \code{OpenDAPOpenDAPYVector} and \code{roiSpatialIndexBound} are optional. They are automatically calculated from the other input parameters if not provided. However, providing them optimizes the performances (i.e. fasten the processing time).
It might be particularly useful to provide them when looping with the function over the same ROI.

Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))}) or as a POSIXlt single time or time range (e.g. "2010-01-01 18:00:00") for the half-hourly collection (GPM_3IMERGHH.06). If POSIXlt, times must be in UTC.
Arguments \code{OpenDAPOpenDAPXVector}, \code{OpenDAPOpenDAPYVector} and \code{roiSpatialIndexBound} are optional. They are automatically calculated from the other input parameters if not provided. However, providing them optimizes the performances (i.e. fasten the processing time).
It might be particularly useful to provide them when looping with the function over the same ROI.
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getData"),quiet=T)
timeRange<-c("2010-01-01 18:00:00","2010-01-02 09:00:00") \%>\% as.POSIXlt()

\dontrun{
getData_era5(timeRange=timeRange,
roi=roi,
dimensions=c("10m_u_component_of_wind","10m_v_component_of_wind"),
username="my.cophub.username",
password="my.cophub.pw",
download=FALSE
)
}


# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getData"),quiet=T)
timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getData_gpm(timeRange=timeRange,
roi=roi,
collection="GPM_3IMERGDF.06",
dimensions=c("precipitationCal"),
username="my.earthdata.username",
password="my.earthdata.pw",
download=FALSE
)
}

}
\seealso{
Other getData: \code{\link{getData_imcce}},
  \code{\link{getData_modis}},
  \code{\link{getData_tamsat}},
  \code{\link{getData_viirsDnb}}

Other getData: \code{\link{getData_imcce}},
  \code{\link{getData_modis}},
  \code{\link{getData_tamsat}},
  \code{\link{getData_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}

Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getData}
