% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_imcce.R
\name{getData_imcce}
\alias{getData_imcce}
\title{Download IMCCE time series data}
\usage{
getData_imcce(timeRange = as.Date(c("2010-01-01", "2010-01-30")),
  roi = st_read(system.file("extdata/ROI_example.kml", package =
  "getRemoteData"), quiet = T), download = FALSE, destFolder = getwd(),
  ...)
}
\arguments{
\item{timeRange}{Date(s) of interest. Mandatory. See Details for addition information on how to provide the dates.}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}
}
\value{
a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

If \code{download} is set to TRUE, the data are downloaded to \code{destFolder}
}
\description{
This function enables to retrieve URLs of IMCCE datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))})
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=T)
timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getData_imcce(timeRange=timeRange,
roi=roi,
collection="GPM_3IMERGDF.06",
dimensions=c("precipitationCal"),
username="my.earthdata.username",
password="my.earthdata.pw",
download=FALSE
)
}

}
\seealso{
Other getData: \code{\link{getData_gpm}},
  \code{\link{getData_modis_vnp}},
  \code{\link{getData_smap}}, \code{\link{getData_srtm}},
  \code{\link{getData_tamsat}},
  \code{\link{getData_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getData}
