% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_modis_vnp.R, R/prepareData_modis_vnp.R
\name{getData_modis_vnp}
\alias{getData_modis_vnp}
\title{Download MODIS or VNP collections time series data}
\usage{
getData_modis_vnp(path_to_raw_modis, var_name)

getData_modis_vnp(path_to_raw_modis, var_name)
}
\arguments{
\item{path_to_raw_modis}{string. Path to the data in .nc4 format}

\item{var_name}{string. Dimension name (e.g. "LST_Day_1km")}

\item{timeRange}{Date(s) of interest. Mandatory. Either a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided with two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))})}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}

\item{collection}{string. The MODIS collection of interest.}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the MODIS collection of interest.}

\item{modisTile}{string. Optional. The MODIS tile name.}

\item{OpenDAPtimeVector}{numeric vector. Optional. The OpenDAP time dimension vector.}

\item{OpenDAPXVector}{numeric vector. Optional. The OpenDAP X (longitude) dimension vector.}

\item{OpenDAPYVector}{numeric vector. Optional. The OpenDAP Y (latitude) dimension vector.}

\item{roiSpatialIndexBound}{numeric vector. Optional.}

\item{username}{string. EarthData username}

\item{password}{string. EarthData password}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}
}
\value{
a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

Additionally, if \code{download} is set to TRUE the data are downloaded to \code{destFolder}

a \code{raster} object with the MODIS projection (+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs)
}
\description{
This function enables to retrieve URLs of MODIS datasets for a given ROI and time frame, and eventually download the data

This function opens as a \code{raster} object a MODIS product that was downloaded via the \code{getData_modis} and \code{dowloadData} functions
}
\details{
Arguments \code{modisTile}, \code{OpenDAPtimeVector}, \code{OpenDAPXVector}, \code{OpenDAPYVector} and \code{roiSpatialIndexBound} are optional. They are automatically calculated from the other input parameters if not provided. However, providing them optimizes the performances (i.e. fasten the processing time).
It might be particularly useful to provide them when looping with the function over the same ROI.

Currently, download of data over multiple MODIS tiles is not enabled.
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=T)
timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getData_modis_vnp(timeRange=timeRange,
roi=roi,
collection="MOD11A1.006",
dimensions=c("LST_Day_1km","LST_Night_1km"),
username="my.earthdata.username",
password="my.earthdata.pw",
download=FALSE
)
}

path_to_modis<-system.file("extdata/modis_example.nc4", package = "getRemoteData")

rast_modis<-getData_modis_vnp(path_to_raw_modis=path_to_modis,
var_name="LST_Day_1km")
#plot(rast_modis)

}
\seealso{
Other getData: \code{\link{getData_gpm}},
  \code{\link{getData_imcce}}, \code{\link{getData_smap}},
  \code{\link{getData_srtm}}, \code{\link{getData_tamsat}},
  \code{\link{getData_viirsDnb}}

Other prepareData: \code{\link{prepareData_era5}},
  \code{\link{prepareData_gpm}},
  \code{\link{prepareData_smap}},
  \code{\link{prepareData_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}

Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getData}
\concept{prepareData}
