% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_srtm.R
\name{getData_srtm}
\alias{getData_srtm}
\title{Download SRTM DEM}
\usage{
getData_srtm(roi = st_read(system.file("extdata/ROI_example.kml", package
  = "getRemoteData"), quiet = T), download = FALSE,
  destFolder = getwd(), ...)
}
\arguments{
\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}
}
\value{
a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

Additionally, if \code{download} is set to TRUE the data are downloaded to \code{destFolder}
}
\description{
This function enables to retrieve URLs of SRTM DEM datasets for a given ROI, and eventually download the data
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=T)

\dontrun{
getData_srtm(roi=roi,
download=FALSE
)
}
}
\seealso{
Other getData: \code{\link{getData_gpm}},
  \code{\link{getData_imcce}}, \code{\link{getData_modis}},
  \code{\link{getData_tamsat}},
  \code{\link{getData_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getData}
