% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_tamsat.R
\name{getData_tamsat}
\alias{getData_tamsat}
\title{Download TAMSAT time series data}
\usage{
getData_tamsat(timeRange = as.Date(c("2017-01-01", "2017-01-30")),
  output_time_step = "daily", output_product = "rainfall_estimate",
  output_output = "individual", download = FALSE,
  destFolder = getwd(), ...)
}
\arguments{
\item{timeRange}{Date(s) of interest. Mandatory. See Details for addition information on how to provide the dates.}

\item{download}{logical. Download data ?}

\item{destFolder}{string. Mandatory if \code{download} is set to TRUE. The destination folder (i.e. folder where the data will be downloaded)}
}
\value{
a data.frame with 3 columns :
 \itemize{
 \item{"name": }{Names (unique identifiers) of each dataset }
 \item{"url": }{URL to download the dataset}
 \item{"destfile": }{Local destination file}
 }

If \code{download} is set to TRUE, the data are downloaded to \code{destFolder}
}
\description{
This function enables to retrieve URLs of TAMSAT datasets for a given time frame, and eventually download the data
}
\details{
Additional information : https://www.tamsat.org.uk/data/archive
}
\examples{

timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getData_tamsat(timeRange=timeRange,
output_time_step="daily",
output_product="rainfall_estimate",
output_output="individual",
download=FALSE
)
}

}
\seealso{
Other getData: \code{\link{getData_gpm}},
  \code{\link{getData_imcce}},
  \code{\link{getData_modis_vnp}},
  \code{\link{getData_smap}}, \code{\link{getData_srtm}},
  \code{\link{getData_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getData}
