% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getremotedata-package.R
\docType{package}
\name{getremotedata-package}
\alias{getremotedata}
\alias{getremotedata-package}
\title{getremotedata: R functions to download & import various
    spatio-temporal data}
\description{
This package contains functions to download
    and pre-process public domain data, eventually containing spatial and
    temporal components.  Particularly, when web-services are implemented
    by the data provider, it enables to download the data for a specific
    region of interest and/or time frame, hence speeding up the overall
    downloading and pre-processing of data on relatively small regions of
    interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ptaconet/getremotedata}
  \item Report bugs at \url{https://github.com/ptaconet/getremotedata/issues}
}

}
\author{
\strong{Maintainer}: Paul Taconet \email{paul.taconet@ird.fr}

}
\keyword{internal}
