% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_era5.R
\name{getUrl_era5}
\alias{getUrl_era5}
\title{Download Copernicus ERA-5 time series data}
\usage{
getUrl_era5(timeRange, roi, dimensions, username = NULL,
  password = NULL, download = FALSE, destFolder = getwd(), ...)
}
\arguments{
\item{timeRange}{date(s) of interest (single date/datetime or time frame) (see details)}

\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the collection of interest}

\item{username}{string. Copernicus username}

\item{password}{string. Copernicus password}
}
\value{
a data.frame with one row for each dataset and 3 columns  :
 \itemize{
 \item{*time_start*: }{Start Date/time for the dataset}
 \item{*name*: }{An indicative name for the dataset}
 \item{*url*: }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of ERA-5 products given a ROI, a time frame and a set of dimensions of interest.
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=TRUE)
timeRange<-c("2010-01-01 18:00:00","2010-01-02 09:00:00") \%>\% as.POSIXlt()

\dontrun{
getUrl_era5(timeRange=timeRange,
roi=roi,
dimensions=c("10m_u_component_of_wind","10m_v_component_of_wind"),
username="my.cophub.username",
password="my.cophub.pw",
download=FALSE
)
}

}
\seealso{
Other getUrl: \code{\link{getUrl_gpm}},
  \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_tamsat}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
