% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_gpm.R
\name{getUrl_gpm}
\alias{getUrl_gpm}
\title{Query and download GPM collections}
\usage{
getUrl_gpm(timeRange, roi, collection, dimensions,
  optionals_opendap = NULL, username = NULL, password = NULL)
}
\arguments{
\item{timeRange}{Date(s) of interest (single date or time frame) (see details)}

\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}

\item{collection}{string. Collection of interest}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the collection of interest}

\item{optionals_opendap}{list of optional arguments (see details). to retrieve outside the function to fasten the processing time}

\item{username}{string. EarthData username}

\item{password}{string. EarthData password}
}
\value{
a data.frame with one row for each dataset and 3 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"name": }{A name for the dataset. Name is indicative}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of GPM datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{optionals_opendap} is optional. This parameter is automatically calculated within the function if it is not provided. However, providing it optimizes the performances of the function (i.e. fasten the processing time).
It might be particularly useful to provide it if looping over the same ROI or dates is planned.
The parameter can be retrieved outside the function with the function \link[getRemoteData]{getOpendapOptArguments_gpm}.

Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))}) or as a POSIXlt single time or time range (e.g. "2010-01-01 18:00:00") for the half-hourly collection (GPM_3IMERGHH.06). If POSIXlt, times must be in UTC.
}
\examples{

require(sf)
require(purrr)

# Set credentials to EarthData servers
my.earthdata.username<-"username"
my.earthdata.pw<-"password"

# Set ROI and time range of interest
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=TRUE)
timeRange<-as.Date(c("2017-01-01","2017-01-30"))

# Retrieve the URLs to download GPM Daily precipitation final run (GPM_3IMERGDF.06) (band precipitationCal and precipitationCal_cnt)
\dontrun{
df_data_to_dl<-getUrl_gpm(
timeRange=timeRange,
roi=roi,
collection="GPM_3IMERGDF.06",
dimensions=c("precipitationCal","precipitationCal_cnt"),
username=my.earthdata.username,
password=my.earthdata.pw
)

# Set destination folder
df_data_to_dl$destfile<-file.path(getwd(),df_data_to_dl$name)

# Download the data
res_dl<-getRemoteData::downloadData(df_data_to_dl,my.earthdata.username,my.earthdata.pw)

# Open the LST_Day_1km bands as a list of rasters
rasts_gpm_day<-purrr::map(res_dl$destfile,~getRemoteData::importData_gpm(.,"precipitationCal")) \%>\%
purrr::set_names(res_dl$name)


}

}
\seealso{
Other getUrl: \code{\link{getUrl_era5}},
  \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_tamsat}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
