% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_imcce.R
\name{getUrl_imcce}
\alias{getUrl_imcce}
\title{Download IMCCE time series data}
\usage{
getUrl_imcce(timeRange, roi)
}
\arguments{
\item{timeRange}{Date(s) of interest (single date or time frame) (see details)}

\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}
}
\value{
a data.frame with one row for each dataset and 2 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of IMCCE datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))})
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=T)
timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getUrl_imcce(timeRange=timeRange,
roi=roi,
collection="GPM_3IMERGDF.06",
dimensions=c("precipitationCal"),
username="my.earthdata.username",
password="my.earthdata.pw",
download=FALSE
)
}

}
\seealso{
Other getUrl: \code{\link{getUrl_era5}},
  \code{\link{getUrl_gpm}}, \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_tamsat}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
