% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_viirsDnb.R
\name{getUrl_viirsDnb}
\alias{getUrl_viirsDnb}
\title{Download VIIRS DNB time series data}
\usage{
getUrl_viirsDnb(timeRange, roi, dimensions)
}
\arguments{
\item{timeRange}{Date(s) of interest (single date or time frame) (see details)}

\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the collection of interest}
}
\value{
a data.frame with one row for each dataset and 2 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of VIIRS DNB datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))})
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=TRUE)
timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

getUrl_viirsDnb(timeRange=timeRange,
roi=roi,
dimensions=c("Monthly_AvgRadiance","Monthly_CloudFreeCoverage"),
download=FALSE
)


}
\seealso{
Other getUrl: \code{\link{getUrl_era5}},
  \code{\link{getUrl_gpm}}, \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
