% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd_import_data.R
\name{grd_import_data}
\alias{grd_import_data}
\title{Import the data in R}
\usage{
grd_import_data(df_data_to_import, collection, variable = NULL,
  roi = NULL, output = "RasterBrick")
}
\arguments{
\item{df_data_to_import}{data.frame. Contains the path to the data that were downloaded. Typically output \code{grd_get_url}}

\item{collection}{character string the name of a collection}

\item{variable}{character string the name of a variables}

\item{roi}{sf the ROI. Mandatory for SMAP collection, else, not necessary}

\item{output}{character string the output. "RasterBrick" .}
}
\description{
Function to import a data / time series that was downloaded with the package \code{getremotedata}
}
