% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd_list_variables.R
\name{grd_list_variables}
\alias{grd_list_variables}
\title{Get informations related to the variables available for a given collection}
\usage{
grd_list_variables(collection)
}
\arguments{
\item{collection}{string. mandatory. Collection of interest (see details).}
}
\value{
A data.frame with the available variables for the collection, and a set of related information for each variable.
}
\description{
Get the variables available for a given collection
}
\examples{

# Get the variables available for the collection VIIRS_DNB_MONTH
(df_varinfo <- grd_list_variables("VIIRS_DNB_MONTH"))


}
