% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData_tamsat.R
\name{importData_tamsat}
\alias{importData_tamsat}
\title{Open a TAMSAT product as a \code{raster} object}
\usage{
importData_tamsat(path_to_raw_tamsat,
  roi = sf::st_read(system.file("extdata/ROI_example.kml", package =
  "getRemoteData"), quiet = T))
}
\arguments{
\item{path_to_raw_tamsat}{string. Path to the data in .nc4 format}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}
}
\value{
a \code{raster} object (EPSG 4326) crop in the ROI.
}
\description{
This function opens as a \code{raster} object a TAMSAT product that was downloaded via the \code{getData_tamsat} and \code{dowloadData} functions
}
\examples{

\dontrun{
path_to_tamsat<-"path.to.my.tamsat.raster"
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=T)
rast_tamsat<-importData_tamsat(path_to_raw_tamsat=path_to_tamsat,
roi=roi)
#plot(rast_tamsat)
}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
