% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_era5.R
\name{prepareData_era5}
\alias{prepareData_era5}
\title{Open a ERA-5 product as a \code{raster} object}
\usage{
prepareData_era5(path_to_raw_era5)
}
\arguments{
\item{path_to_raw_modis}{string. Path to the data in .nc format}

\item{var_name}{string. Dimension name (e.g. "LST_Day_1km")}
}
\value{
a \code{raster} object with the MODIS projection (+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs)
}
\description{
This function opens as a \code{raster} object a ERA-5 product that was downloaded via the \code{getData_era5} function
}
\examples{

path_to_era5<-system.file("extdata/era5_example.nc", package = "getRemoteData")

rast_era5<-prepareData_modis(path_to_raw_era5=path_to_era5)
#plot(rast_era5)

}
\seealso{
Other prepareData: \code{\link{prepareData_gpm}},
  \code{\link{prepareData_modis}},
  \code{\link{prepareData_smap}},
  \code{\link{prepareData_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
