% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_gpm.R
\name{prepareData_gpm}
\alias{prepareData_gpm}
\title{Open a GPM product as a \code{raster} object}
\usage{
prepareData_gpm(path_to_raw_gpm, var_name)
}
\arguments{
\item{var_name}{string. Dimension name (e.g. "precipitationCal")}

\item{path_to_raw_modis}{string. Path to the data in .nc4 format}
}
\value{
a \code{raster} object (EPSG 4326)
}
\description{
This function opens as a \code{raster} object a GPM product that was downloaded via the \code{getData_modis} and \code{dowloadData} functions
}
\examples{

path_to_gpm<-system.file("extdata/gpm_example.nc4", package = "getRemoteData")

rast_gpm<-prepareData_gpm(path_to_raw_gpm=path_to_gpm,
var_name="precipitationCal")
#plot(rast_modis)

}
\seealso{
Other prepareData: \code{\link{prepareData_era5}},
  \code{\link{prepareData_modis}},
  \code{\link{prepareData_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
