% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_modis.R
\name{prepareData_modis}
\alias{prepareData_modis}
\title{Open a MODIS product as a \code{raster} object}
\usage{
prepareData_modis(path_to_raw_modis, var_name)
}
\arguments{
\item{path_to_raw_modis}{string. Path to the data in .nc4 format}

\item{var_name}{string. Dimension name (e.g. "LST_Day_1km")}
}
\value{
a \code{raster} object with the MODIS projection (+proj=sinu +lon_0=0 +x_0=0 +y_0=0 +a=6371007.181 +b=6371007.181 +units=m +no_defs)
}
\description{
This function opens as a \code{raster} object a MODIS product that was downloaded via the \code{getData_modis} and \code{dowloadData} functions
}
\examples{

path_to_modis<-system.file("extdata/modis_example.nc4", package = "getRemoteData")

rast_modis<-prepareData_modis(path_to_raw_modis=path_to_modis,
var_name="LST_Day_1km")
#plot(rast_modis)

}
\seealso{
Other prepareData: \code{\link{prepareData_gpm}},
  \code{\link{prepareData_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
