% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_smap.R
\name{prepareData_smap}
\alias{prepareData_smap}
\title{Open a SMAP product as a \code{raster} object}
\usage{
prepareData_smap(path_to_raw_smap, var_name, minLon, minLat, maxLon,
  maxLat)
}
\arguments{
\item{var_name}{string. Dimension name (e.g. "Soil_Moisture_Retrieval_Data_AM_soil_moisture")}

\item{path_to_raw_modis}{string. Path to the data in .nc4 format}
}
\value{
a \code{raster} object (EPSG 4326)
}
\description{
This function opens as a \code{raster} object a SMAP product that was downloaded via the \code{getData_smap} and \code{dowloadData} functions
}
\examples{

path_to_gpm<-system.file("extdata/smap_example.nc4", package = "getRemoteData")

rast_smap<-prepareData_smap(path_to_raw_smap=path_to_smap,
var_name="Soil_Moisture_Retrieval_Data_AM_soil_moisture")
#plot(rast_smap)

}
\seealso{
Other prepareData: \code{\link{prepareData_era5}},
  \code{\link{prepareData_gpm}},
  \code{\link{prepareData_modis}},
  \code{\link{prepareData_tamsat}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
