% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_tamsat.R
\name{prepareData_tamsat}
\alias{prepareData_tamsat}
\title{Open a TAMSAT product as a \code{raster} object}
\usage{
prepareData_tamsat(path_to_raw_tamsat,
  roi = st_read("/home/ptaconet/r_react/getData/ROI_test.kml", quiet =
  T))
}
\arguments{
\item{path_to_raw_tamsat}{string. Path to the data in .nc4 format}

\item{roi}{sf POINT or POLYGON. The region of interest in EPSG 4326}
}
\value{
a \code{raster} object (EPSG 4326) crop in the ROI.
}
\description{
This function opens as a \code{raster} object a TAMSAT product that was downloaded via the \code{getData_tamsat} and \code{dowloadData} functions
}
\examples{

\dontrun{
path_to_tamsat<-"path.to.my.tamsat.raster"
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getData"),quiet=T)
rast_tamsat<-prepareData_tamsat(path_to_raw_tamsat=path_to_tamsat,
roi=roi)
#plot(rast_tamsat)
}
}
\seealso{
Other prepareData: \code{\link{prepareData_era5}},
  \code{\link{prepareData_gpm}},
  \code{\link{prepareData_modis}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{prepareData}
