% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadData}
\alias{downloadData}
\title{Wrapper to download several datasets}
\usage{
downloadData(df_to_dl, parallelDL = FALSE, username = NULL,
  password = NULL, data_source = NULL)
}
\description{
This function enables to download datasets, enventually parallelizing the download.
}
