% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvailableDataSources.R
\name{getAvailableDataSources}
\alias{getAvailableDataSources}
\title{Get data sources / collections implemented in the package}
\usage{
getAvailableDataSources()
}
\arguments{
\item{detailed}{Boolean. Return detailed information on each available data collection available ? TRUE by default. FALSE returns a summarized version of the available sources of data.}
}
\value{
a data.frame with the data sources / collections dealt by getData, along with details on each data collection
}
\description{
This function returns a table of the data sources / collections that can be downloaded and pre-processed using the getData package
}
\details{
The output data.frame columns are:
\itemize{
\item{"source": }{The source acronym}
\item{"collection": }{The collection}
\item{"...": }{...TODO : finish description}
}
}
\examples{

sources<-getAvailableDataSources()
sources

}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
