% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRemoteData-package.R
\docType{package}
\name{getRemoteData-package}
\alias{getRemoteData}
\alias{getRemoteData-package}
\title{getRemoteData: Download and pre-process multiple kind of data}
\description{
This package contains functions to download
    and pre-process public domain data, eventually containing spatial and
    temporal components.  Particularly, when web-services are implemented
    by the data provider, it enables to download the data for a specific
    region of interest and/or time frame, hence speeding up the overall
    downloading and pre-processing of data on relatively small regions of
    interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ptaconet/getData}
  \item Report bugs at \url{https://github.com/ptaconet/getData/issues}
}

}
\author{
\strong{Maintainer}: Paul Taconet \email{paul.taconet@ird.fr}

}
\keyword{internal}
