% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_era5.R
\name{getUrl_era5}
\alias{getUrl_era5}
\title{Download Global Precipitation Measurement time series data}
\usage{
getUrl_era5(timeRange, roi, dimensions, username = NULL,
  password = NULL, download = FALSE, destFolder = getwd(), ...)
}
\arguments{
\item{timeRange}{Date(s) of interest (single date or time frame) (see details)}

\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}

\item{dimensions}{string vector. Names of the dimensions to retrieve for the collection of interest}

\item{username}{string. EarthData username}

\item{password}{string. EarthData password}
}
\value{
a data.frame with one row for each dataset and 2 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of GPM datasets for a given ROI and time frame, and eventually download the data
}
\details{
Argument \code{timeRange} can be provided either as a single date (e.g. \code{as.Date("2017-01-01"))} or time frame provided as two bounding dates ( e.g. \code{as.Date(c("2010-01-01","2010-01-30"))}) or as a POSIXlt single time or time range (e.g. "2010-01-01 18:00:00") for the half-hourly collection (GPM_3IMERGHH.06). If POSIXlt, times must be in UTC.
Arguments \code{OpenDAPOpenDAPXVector}, \code{OpenDAPOpenDAPYVector} and \code{roiSpatialIndexBound} are optional. They are automatically calculated from the other input parameters if not provided. However, providing them optimizes the performances (i.e. fasten the processing time).
It might be particularly useful to provide them when looping with the function over the same ROI.
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=TRUE)
timeRange<-c("2010-01-01 18:00:00","2010-01-02 09:00:00") \%>\% as.POSIXlt()

\dontrun{
getUrl_era5(timeRange=timeRange,
roi=roi,
dimensions=c("10m_u_component_of_wind","10m_v_component_of_wind"),
username="my.cophub.username",
password="my.cophub.pw",
download=FALSE
)
}

}
\seealso{
Other getUrl: \code{\link{getUrl_gpm}},
  \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_tamsat}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
