% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_srtm.R
\name{getUrl_srtm}
\alias{getUrl_srtm}
\title{Download SRTM DEM}
\usage{
getUrl_srtm(roi)
}
\arguments{
\item{roi}{sf POINT or POLYGON. Region of interest (EPSG 4326)}
}
\value{
a data.frame with one row for each dataset and 2 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of SRTM DEM datasets for a given ROI, and eventually download the data
}
\examples{

# Read ROI as sf object
roi=sf::st_read(system.file("extdata/ROI_example.kml", package = "getRemoteData"),quiet=TRUE)

\dontrun{
getUrl_srtm(roi=roi,
download=FALSE
)
}
}
\seealso{
Other getUrl: \code{\link{getUrl_era5}},
  \code{\link{getUrl_gpm}}, \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_tamsat}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
