% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrl_tamsat.R
\name{getUrl_tamsat}
\alias{getUrl_tamsat}
\title{Download TAMSAT time series data}
\usage{
getUrl_tamsat(timeRange, output_time_step, output_product, output_output)
}
\arguments{
\item{timeRange}{Date(s) of interest (single date or time frame) (see details)}

\item{output_time_step}{blabla}

\item{output_product}{blibli}

\item{output_output}{bloblo}
}
\value{
a data.frame with one row for each dataset and 3 columns  :
 \itemize{
 \item{"time_start": }{Date or time start}
 \item{"name": }{A name for the dataset. Name is indicative}
 \item{"url": }{URL of the dataset}
 }
}
\description{
This function enables to retrieve URLs of TAMSAT datasets for a given time frame, and eventually download the data
}
\details{
Additional information : https://www.tamsat.org.uk/data/archive
}
\examples{

timeRange<-c("2017-01-01","2017-01-30") \%>\% as.Date()

\dontrun{
getUrl_tamsat(timeRange=timeRange,
output_time_step="daily",
output_product="rainfall_estimate",
output_output="individual",
download=FALSE
)
}

}
\seealso{
Other getUrl: \code{\link{getUrl_era5}},
  \code{\link{getUrl_gpm}}, \code{\link{getUrl_imcce}},
  \code{\link{getUrl_modis_vnp}},
  \code{\link{getUrl_smap}}, \code{\link{getUrl_srtm}},
  \code{\link{getUrl_viirsDnb}}
}
\author{
Paul Taconet, IRD \email{paul.taconet@ird.fr}
}
\concept{getUrl}
