% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd_list_collections.R
\name{grd_list_collections}
\alias{grd_list_collections}
\title{Get data sources / collections implemented in the package}
\usage{
grd_list_collections()
}
\value{
a data.frame with the data sources / collections dealt by grd_list_collections, along with details on each data collection
}
\description{
This function returns a table of the data sources / collections that can be downloaded and pre-processed using the getremotedata package
}
\examples{

sources<-grd_list_collections()
sources

}
\author{
Paul Taconet, \email{paul.taconet@ird.fr}
}
